+++ WatsonTcp/ClientMetadata.cs
@@ -76,14 +76,12 @@ namespace WatsonTcp
             if (_SslStream != null)
             {
                 _SslStream.Close();
-                _SslStream.Dispose();
                 _SslStream = null;
             }
 
             if (_NetworkStream != null)
             {
                 _NetworkStream.Close();
-                _NetworkStream.Dispose();
                 _NetworkStream = null;
             }
 
@@ -109,7 +107,6 @@ namespace WatsonTcp
             if (_TcpClient != null)
             {
                 _TcpClient.Close();
-                _TcpClient.Dispose();
                 _TcpClient = null;
             }
         }